Move py3status code to a separate file
